refactor deck layers code
