+++ src/components/controlPanel.tsx
@@ -4,17 +4,17 @@ import CITIES from '../components/map/data/cities.json';
 
 function ControlPanel({ onSelectCity }) {
   return (
-    <div className="control-panel">
-      <h3>Camera Transition</h3>
+    <div className="absolute top-0 right-0 m-3">
+      <h3>Camera Transition is not finished</h3>
       <hr />
 
       {CITIES.filter(city => city.state === 'California').map((city, index) => (
         <div key={`btn-${index}`} className="input">
           <input
-            type="radio"
-            name="city"
-            id={`city-${index}`}
             defaultChecked={city.city === 'San Francisco'}
+            id={`city-${index}`}
+            name="city"
+            type="radio"
             onClick={() => onSelectCity(city)}
           />
           <label htmlFor={`city-${index}`}>Hello</label>
