--- src/components/map/constatnts.tsx
@@ -1,3 +1,0 @@
-export const MAP_BOX_TOKEN =
-  process.env.REACT_APP_MAP_BOX_TOKEN ||
-  'pk.eyJ1Ijoic2xhdzY2NiIsImEiOiJjbGF6dWs5MXAwYng1M25wbDNpNjlscTdlIn0.iXtSCUNohNGfnRy2SoKRvQ';
