+++ src/components/map/constats.tsx
@@ -1,0 +1,12 @@
+export const MAP_BOX_TOKEN =
+  'pk.eyJ1Ijoic2xhdzY2NiIsImEiOiJjbGF6dWs5MXAwYng1M25wbDNpNjlscTdlIn0.iXtSCUNohNGfnRy2SoKRvQ';
+
+export const MAP_STYLE = 'https://basemaps.cartocdn.com/gl/dark-matter-nolabels-gl-style/style.json';
+
+export const INITIAL_VIEW_STATE = {
+  latitude: 45.47,
+  longitude: -75.45,
+  zoom: 10,
+  bearing: 0,
+  pitch: 10,
+};
